backend/pyproject.toml
[project]
name = "medtimeline-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.8"
dependencies = [
    "alembic>=1.18.4",
    "arq>=0.27.0",
    "asyncpg>=0.31.0",
    "bcrypt<5",
    "email-validator>=2.3.0",
    "fastapi>=0.128.8",
    "fhir-resources>=8.2.0",
    "fhirpathpy>=2.1.0",
    "httpx>=0.28.1",
    "ijson>=3.4.0.post0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn[standard]>=0.40.0",
    "google-genai>=1.33.0",
    "langextract>=1.0.7",
    "striprtf>=0.0.28",
    "Pillow>=11.0.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
